+++ src/services/cognito.js
@@ -156,8 +156,7 @@ function getUserProfile() /* : {
     fullName: ?string,
     email: ?string,
     providerUserId: ?string,
-  },
-  preferred_username: ?string,
+  }
 } | null */ {
   if (!awsCognitoClient) {
     return null
@@ -182,7 +181,6 @@ function getUserProfile() /* : {
     role: userProfile['custom:role'],
     username: userProfile.email,
     supervisor: undefined,
-    preferred_username: undefined
   }
 
   if (
@@ -202,8 +200,7 @@ function getUserProfile() /* : {
     user.providerUserId = userProfile.identities[0].userId
     user.isSAMLUser = user.providerType === 'SAML'
     if (user.isSAMLUser) {
-      user.username = user.providerUserId
-      user.preferred_username = userProfile.preferred_username
+      user.username = userProfile.preferred_username || user.providerUserId
     }
   }