+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - [`authService.loginUsernamePassword()`](./docs/auth-service.md#loginusernamepassword)
 - [`authService.changePassword()`](./docs/auth-service.md#changepassword)
 - [`authService.forgotPassword()`](./docs/auth-service.md#forgotpassword)
+- [`authService.registerAuthListener()`](./docs/auth-service.md#registerauthlistener)
 
 ### Removed